Restructure StateBTreeSet/Map into a separate module
