+++ concordium-std/src/impls.rs
@@ -8,7 +8,7 @@ use crate::{
     mem::{self, MaybeUninit},
     num,
     num::NonZeroU32,
-    prims,
+    prims, state_btree,
     traits::*,
     types::*,
     vec::Vec,
@@ -774,8 +774,6 @@ where
 /// "allocator"/state builder stores "next location". The values stored at this
 /// location are 64-bit integers.
 const NEXT_ITEM_PREFIX_KEY: [u8; 8] = 0u64.to_le_bytes();
-#[cfg(test)]
-const GENERIC_MAP_PREFIX: u64 = 1;
 /// Initial location to store in [NEXT_ITEM_PREFIX_KEY]. For example, the
 /// initial call to "new_state_box" will allocate the box at this location.
 pub(crate) const INITIAL_NEXT_ITEM_PREFIX: [u8; 8] = 2u64.to_le_bytes();
@@ -2334,14 +2332,14 @@ where
     }
 
     /// Create a new empty [`StateBTreeSet`].
-    pub fn new_btree_set<K>(&mut self) -> StateBTreeSet<K, S> {
+    pub fn new_btree_set<K>(&mut self) -> state_btree::StateBTreeSet<K, S> {
         let (state_api, prefix) = self.new_state_container();
-        StateBTreeSet::new(state_api, prefix)
+        state_btree::StateBTreeSet::new(state_api, prefix)
     }
 
     /// Create a new empty [`StateBTreeMap`].
-    pub fn new_btree_map<K, V>(&mut self) -> StateBTreeMap<K, V, S> {
-        StateBTreeMap {
+    pub fn new_btree_map<K, V>(&mut self) -> state_btree::StateBTreeMap<K, V, S> {
+        state_btree::StateBTreeMap {
             map:         self.new_map(),
             ordered_set: self.new_btree_set(),
         }
@@ -2349,15 +2347,19 @@ where
 
     /// Create a new empty [`StateBTreeSet`], setting the minimum degree `M` of
     /// the B-Tree explicitly.
-    pub fn new_btree_set_degree<const M: usize, K>(&mut self) -> StateBTreeSet<K, S, M> {
+    pub fn new_btree_set_degree<const M: usize, K>(
+        &mut self,
+    ) -> state_btree::StateBTreeSet<K, S, M> {
         let (state_api, prefix) = self.new_state_container();
-        StateBTreeSet::new(state_api, prefix)
+        state_btree::StateBTreeSet::new(state_api, prefix)
     }
 
     /// Create a new empty [`StateBTreeMap`], setting the minimum degree `M` of
     /// the B-Tree explicitly.
-    pub fn new_btree_map_degree<const M: usize, K, V>(&mut self) -> StateBTreeMap<K, V, S, M> {
-        StateBTreeMap {
+    pub fn new_btree_map_degree<const M: usize, K, V>(
+        &mut self,
+    ) -> state_btree::StateBTreeMap<K, V, S, M> {
+        state_btree::StateBTreeMap {
             map:         self.new_map(),
             ordered_set: self.new_btree_set_degree(),
         }
@@ -2435,51 +2437,6 @@ where
         next_collection_prefix_entry.write_u64(collection_prefix + 1).unwrap_abort(); // Writing to state cannot fail.
 
         collection_prefix.to_le_bytes()
-    }
-}
-
-#[cfg(test)]
-/// Some helper methods that are used for internal tests.
-impl<S> StateBuilder<S>
-where
-    S: HasStateApi,
-{
-    /// Get a value from the generic map.
-    /// `Some(Err(_))` means that something exists in the state with that key,
-    /// but it isn't of type `V`.
-    pub(crate) fn get<K: Serial, V: DeserialWithState<S>>(&self, key: K) -> Option<ParseResult<V>> {
-        let key_with_map_prefix = Self::prepend_generic_map_key(key);
-
-        self.state_api
-            .lookup_entry(&key_with_map_prefix)
-            .map(|mut entry| V::deserial_with_state(&self.state_api, &mut entry))
-    }
-
-    /// Inserts a value in the generic map.
-    /// The key and value are serialized before insert.
-    pub(crate) fn insert<K: Serial, V: Serial>(
-        &mut self,
-        key: K,
-        value: V,
-    ) -> Result<(), StateError> {
-        let key_with_map_prefix = Self::prepend_generic_map_key(key);
-        match self.state_api.entry(key_with_map_prefix) {
-            EntryRaw::Vacant(vac) => {
-                let _ = vac.insert(&value);
-            }
-            EntryRaw::Occupied(mut occ) => occ.insert(&value),
-        }
-        Ok(())
-    }
-
-    /// Serializes the key and prepends [GENERIC_MAP_PREFIX].
-    /// This is similar to how [StateMap] works, where a unique prefix is
-    /// prepended onto keys. Since there is only one generic map, the prefix
-    /// is a constant.
-    fn prepend_generic_map_key<K: Serial>(key: K) -> Vec<u8> {
-        let mut key_with_map_prefix = to_bytes(&GENERIC_MAP_PREFIX);
-        key_with_map_prefix.append(&mut to_bytes(&key));
-        key_with_map_prefix
     }
 }
 
@@ -2612,7 +2569,6 @@ where
         (&mut self.state, &mut self.state_builder)
     }
 }
-
 impl HasHost<ExternStateApi> for ExternLowLevelHost {
     type ReturnValueType = ExternCallResponse;
     type StateApiType = ExternStateApi;
@@ -3179,1005 +3135,6 @@ impl Deserial for MetadataUrl {
     }
 }
 
-impl<const M: usize, K, V, S> Serial for StateBTreeMap<K, V, S, M> {
-    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
-        self.map.serial(out)?;
-        self.ordered_set.serial(out)
-    }
-}
-
-impl<const M: usize, K, V, S: HasStateApi> DeserialWithState<S> for StateBTreeMap<K, V, S, M> {
-    fn deserial_with_state<R: Read>(state: &S, source: &mut R) -> ParseResult<Self> {
-        let map = DeserialWithState::deserial_with_state(state, source)?;
-        let ordered_set = DeserialWithState::deserial_with_state(state, source)?;
-        Ok(Self {
-            map,
-            ordered_set,
-        })
-    }
-}
-
-impl<const M: usize, K, S> Serial for StateBTreeSet<K, S, M> {
-    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
-        self.prefix.serial(out)?;
-        self.root.serial(out)?;
-        self.len.serial(out)?;
-        self.next_node_id.serial(out)
-    }
-}
-
-impl<const M: usize, K, S: HasStateApi> DeserialWithState<S> for StateBTreeSet<K, S, M> {
-    fn deserial_with_state<R: Read>(state: &S, source: &mut R) -> ParseResult<Self> {
-        let prefix = source.get()?;
-        let root = source.get()?;
-        let len = source.get()?;
-        let next_node_id = source.get()?;
-
-        Ok(Self {
-            _marker_key: Default::default(),
-            prefix,
-            state_api: state.clone(),
-            root,
-            len,
-            next_node_id,
-        })
-    }
-}
-
-impl<const M: usize, K: Serial> Serial for state_btree_internals::Node<M, K> {
-    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> {
-        self.keys.serial(out)?;
-        self.children.serial(out)
-    }
-}
-
-impl<const M: usize, K: Deserial> Deserial for state_btree_internals::Node<M, K> {
-    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
-        let keys = source.get()?;
-        let children = source.get()?;
-        Ok(Self {
-            keys,
-            children,
-        })
-    }
-}
-
-impl Serial for state_btree_internals::NodeId {
-    fn serial<W: Write>(&self, out: &mut W) -> Result<(), W::Err> { self.id.serial(out) }
-}
-
-impl Deserial for state_btree_internals::NodeId {
-    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
-        Ok(Self {
-            id: source.get()?,
-        })
-    }
-}
-
-impl<const M: usize, K, V, S> StateBTreeMap<K, V, S, M> {
-    /// Insert a key-value pair into the map.
-    /// Returns the previous value if the key was already in the map.
-    pub fn insert(&mut self, key: K, value: V) -> Option<V>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord,
-        V: Serial + DeserialWithState<S>, {
-        let old_value_option = self.map.insert_borrowed(&key, value);
-        if old_value_option.is_none() && !self.ordered_set.insert(key) {
-            // Inconsistency between the map and ordered_set.
-            crate::trap();
-        }
-        old_value_option
-    }
-
-    /// Remove a key from the map, returning the value at the key if the key was
-    /// previously in the map.
-    ///
-    /// *Caution*: If `V` is a [StateBox], [StateMap], then it is
-    /// important to call [`Deletable::delete`] on the value returned when
-    /// you're finished with it. Otherwise, it will remain in the contract
-    /// state.
-    #[must_use]
-    pub fn remove_and_get(&mut self, key: &K) -> Option<V>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord,
-        V: Serial + DeserialWithState<S> + Deletable, {
-        let v = self.map.remove_and_get(key);
-        if v.is_some() && !self.ordered_set.remove(key) {
-            // Inconsistency between the map and ordered_set.
-            crate::trap();
-        }
-        v
-    }
-
-    /// Remove a key from the map.
-    /// This also deletes the value in the state.
-    pub fn remove(&mut self, key: &K)
-    where
-        S: HasStateApi,
-        K: Serialize + Ord,
-        V: Serial + DeserialWithState<S> + Deletable, {
-        if self.ordered_set.remove(key) {
-            self.map.remove(key);
-        }
-    }
-
-    /// Get a reference to the value corresponding to the key.
-    pub fn get(&self, key: &K) -> Option<StateRef<V>>
-    where
-        K: Serialize,
-        S: HasStateApi,
-        V: Serial + DeserialWithState<S>, {
-        if self.ordered_set.is_empty() {
-            None
-        } else {
-            self.map.get(key)
-        }
-    }
-
-    /// Get a mutable reference to the value corresponding to the key.
-    pub fn get_mut(&mut self, key: &K) -> Option<StateRefMut<V, S>>
-    where
-        K: Serialize,
-        S: HasStateApi,
-        V: Serial + DeserialWithState<S>, {
-        if self.ordered_set.is_empty() {
-            None
-        } else {
-            self.map.get_mut(key)
-        }
-    }
-
-    /// Returns `true` if the map contains a value for the specified key.
-    pub fn contains_key(&self, key: &K) -> bool
-    where
-        K: Serialize + Ord,
-        S: HasStateApi, {
-        self.ordered_set.contains(key)
-    }
-
-    /// Returns the smallest key in the map, which is strictly larger than the
-    /// provided key. `None` meaning no such key is present in the map.
-    pub fn higher(&self, key: &K) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        self.ordered_set.higher(key)
-    }
-
-    /// Returns the largest key in the map, which is strictly smaller than the
-    /// provided key. `None` meaning no such key is present in the map.
-    pub fn lower(&self, key: &K) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        self.ordered_set.lower(key)
-    }
-
-    /// Returns a reference to the first key in the map, if any. This key is
-    /// always the minimum of all keys in the map.
-    pub fn first_key(&self) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        self.ordered_set.first()
-    }
-
-    /// Returns a reference to the last key in the map, if any. This key is
-    /// always the maximum of all keys in the map.
-    pub fn last_key(&self) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        self.ordered_set.last()
-    }
-
-    /// Return the number of elements in the map.
-    pub fn len(&self) -> u32 { self.ordered_set.len() }
-
-    /// Returns `true` is the map contains no elements.
-    pub fn is_empty(&self) -> bool { self.ordered_set.is_empty() }
-
-    /// Create an iterator over the entries of [`StateBTreeMap`].
-    /// Ordered by `K`.
-    pub fn iter(&self) -> StateBTreeMapIter<K, V, S, M>
-    where
-        S: HasStateApi, {
-        StateBTreeMapIter {
-            key_iter: self.ordered_set.iter(),
-            map:      &self.map,
-        }
-    }
-
-    /// Clears the map, removing all key-value pairs.
-    /// This also includes values pointed at, if `V`, for example, is a
-    /// [StateBox]. **If applicable use [`clear_flat`](Self::clear_flat)
-    /// instead.**
-    pub fn clear(&mut self)
-    where
-        S: HasStateApi,
-        K: Serialize,
-        V: Serial + DeserialWithState<S> + Deletable, {
-        self.map.clear();
-        self.ordered_set.clear();
-    }
-
-    /// Clears the map, removing all key-value pairs.
-    /// **This should be used over [`clear`](Self::clear) if it is
-    /// applicable.** It avoids recursive deletion of values since the
-    /// values are required to be _flat_.
-    ///
-    /// Unfortunately it is not possible to automatically choose between these
-    /// implementations. Once Rust gets trait specialization then this might
-    /// be possible.
-    pub fn clear_flat(&mut self)
-    where
-        S: HasStateApi,
-        K: Serialize,
-        V: Serialize, {
-        self.map.clear_flat();
-        self.ordered_set.clear();
-    }
-}
-
-impl<const M: usize, K, S> StateBTreeSet<K, S, M> {
-    /// Construct a new [`StateBTreeSet`] given a unique prefix to use in the
-    /// key-value store.
-    pub(crate) fn new(state_api: S, prefix: StateItemPrefix) -> Self {
-        Self {
-            _marker_key: Default::default(),
-            prefix,
-            state_api,
-            root: None,
-            len: 0,
-            next_node_id: state_btree_internals::NodeId {
-                id: 0,
-            },
-        }
-    }
-
-    /// Insert a key into the set.
-    /// Returns true if the key is new in the collection.
-    pub fn insert(&mut self, key: K) -> bool
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        let Some(root_id) = self.root else {
-            let node_id = {
-                let (node_id, _node) = self.create_node(crate::vec![key], Vec::new());
-                node_id
-            };
-            self.root = Some(node_id);
-            self.len = 1;
-            return true;
-        };
-
-        let root_node = self.get_node_mut(root_id);
-        if !root_node.is_full() {
-            let new = self.insert_non_full(root_node, key);
-            if new {
-                self.len += 1;
-            }
-            return new;
-        } else if root_node.keys.binary_search(&key).is_ok() {
-            return false;
-        }
-        // The root node is full, so we construct a new root node.
-        let (new_root_id, mut new_root) = self.create_node(Vec::new(), crate::vec![root_id]);
-        self.root = Some(new_root_id);
-        // The old root node is now a child node.
-        let mut child = root_node;
-        let new_larger_child = self.split_child(&mut new_root, 0, &mut child);
-        // new_root should now contain one key and two children, so we need to know
-        // which one to insert into.
-        let child = if new_root.keys[0] < key {
-            new_larger_child
-        } else {
-            child
-        };
-        let new = self.insert_non_full(child, key);
-        if new {
-            self.len += 1;
-        }
-        new
-    }
-
-    /// Returns `true` if the set contains an element equal to the key.
-    pub fn contains(&self, key: &K) -> bool
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        let Some(root_node_id) = self.root else {
-            return false;
-        };
-        let mut node = self.get_node(root_node_id);
-        loop {
-            let Err(child_index) = node.keys.binary_search(key) else {
-                return true;
-            };
-            if node.is_leaf() {
-                return false;
-            }
-            let child_node_id = node.children[child_index];
-            node = self.get_node(child_node_id);
-        }
-    }
-
-    /// Return the number of elements in the set.
-    pub fn len(&self) -> u32 { self.len }
-
-    /// Returns `true` is the set contains no elements.
-    pub fn is_empty(&self) -> bool { self.root.is_none() }
-
-    /// Get an iterator over the elements in the `StateBTreeSet`. The iterator
-    /// returns elements in increasing order.
-    pub fn iter(&self) -> StateBTreeSetIter<K, S, M>
-    where
-        S: HasStateApi, {
-        StateBTreeSetIter {
-            length:            self.len.try_into().unwrap_abort(),
-            next_node:         self.root,
-            depth_first_stack: Vec::new(),
-            tree:              self,
-            _marker_lifetime:  Default::default(),
-        }
-    }
-
-    /// Clears the set, removing all elements.
-    pub fn clear(&mut self)
-    where
-        S: HasStateApi, {
-        // Reset the information.
-        self.root = None;
-        self.next_node_id = state_btree_internals::NodeId {
-            id: 0,
-        };
-        self.len = 0;
-        // Then delete every node store in the state.
-        // Unwrapping is safe when only using the high-level API.
-        self.state_api.delete_prefix(&self.prefix).unwrap_abort();
-    }
-
-    /// Returns the smallest key in the set, which is strictly larger than the
-    /// provided key. `None` meaning no such key is present in the set.
-    pub fn higher(&self, key: &K) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        let Some(root_node_id) = self.root else {
-            return None;
-        };
-
-        let mut node = self.get_node(root_node_id);
-        let mut higher_so_far = None;
-        loop {
-            let higher_key_index = match node.keys.binary_search(key) {
-                Ok(index) => index + 1,
-                Err(index) => index,
-            };
-
-            if node.is_leaf() {
-                return if higher_key_index < node.keys.len() {
-                    // This does not mutate the node in the end, just the representation in memory
-                    // which is freed after the call.
-                    Some(StateRef::new(node.keys.swap_remove(higher_key_index)))
-                } else {
-                    higher_so_far
-                };
-            } else {
-                if higher_key_index < node.keys.len() {
-                    // This does not mutate the node in the end, just the representation in memory
-                    // which is freed after the call.
-                    higher_so_far = Some(StateRef::new(node.keys.swap_remove(higher_key_index)))
-                }
-
-                let child_node_id = node.children[higher_key_index];
-                node = self.get_node(child_node_id);
-            }
-        }
-    }
-
-    /// Returns the largest key in the set, which is strictly smaller than the
-    /// provided key. `None` meaning no such key is present in the set.
-    pub fn lower(&self, key: &K) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        let Some(root_node_id) = self.root else {
-            return None;
-        };
-
-        let mut node = self.get_node(root_node_id);
-        let mut lower_so_far = None;
-        loop {
-            let lower_key_index = match node.keys.binary_search(key) {
-                Ok(index) => index,
-                Err(index) => index,
-            };
-
-            if node.is_leaf() {
-                return if lower_key_index == 0 {
-                    lower_so_far
-                } else {
-                    // lower_key_index cannot be 0 in this case, since the binary search will only
-                    // return 0 in the true branch above.
-                    Some(StateRef::new(node.keys.swap_remove(lower_key_index - 1)))
-                };
-            } else {
-                if lower_key_index > 0 {
-                    // This does not mutate the node in the end, just the representation in memory
-                    // which is freed after the call.
-                    lower_so_far = Some(StateRef::new(node.keys.swap_remove(lower_key_index - 1)));
-                }
-                let child_node_id = node.children[lower_key_index];
-                node = self.get_node(child_node_id)
-            }
-        }
-    }
-
-    /// Returns a reference to the first key in the set, if any. This key is
-    /// always the minimum of all keys in the set.
-    pub fn first(&self) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        let Some(root_node_id) = self.root else {
-            return None;
-        };
-        let mut root = self.get_node(root_node_id);
-        if root.is_leaf() {
-            Some(StateRef::new(root.keys.swap_remove(0)))
-        } else {
-            Some(StateRef::new(self.get_lowest_key(&root, 0)))
-        }
-    }
-
-    /// Returns a reference to the last key in the set, if any. This key is
-    /// always the maximum of all keys in the set.
-    pub fn last(&self) -> Option<StateRef<K>>
-    where
-        S: HasStateApi,
-        K: Serialize + Ord, {
-        let Some(root_node_id) = self.root else {
-            return None;
-        };
-        let mut root = self.get_node(root_node_id);
-        if root.is_leaf() {
-            Some(StateRef::new(root.keys.pop().unwrap_abort()))
-        } else {
-            Some(StateRef::new(self.get_highest_key(&root, root.children.len() - 1)))
-        }
-    }
-
-    /// Remove a key from the set.
-    /// Returns whether such an element was present.
-    pub fn remove(&mut self, key: &K) -> bool
-    where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        let Some(root_node_id) = self.root else {
-            return false;
-        };
-
-        let deleted_something = {
-            let mut node = self.get_node_mut(root_node_id);
-            loop {
-                match node.keys.binary_search(key) {
-                    Ok(index) => {
-                        if node.is_leaf() {
-                            // Found the key in this node and the node is a leaf, meaning we
-                            // simply remove it.
-                            // This will not violate the minimum keys invariant, since a node
-                            // ensures a child can spare a key before iteration and the root
-                            // node is not part of the
-                            // invariant.
-                            node.keys.remove(index);
-                            break true;
-                        }
-                        // Found the key in this node, but the node is not a leaf.
-                        let mut left_child = self.get_node_mut(node.children[index]);
-                        if !left_child.is_at_min() {
-                            // If the child with smaller keys can spare a key, we take the
-                            // highest key from it.
-                            node.keys[index] = self.remove_largest_key(left_child);
-                            break true;
-                        }
-
-                        let right_child = self.get_node_mut(node.children[index + 1]);
-                        if !right_child.is_at_min() {
-                            // If the child with larger keys can spare a key, we take the lowest
-                            // key from it.
-                            node.keys[index] = self.remove_smallest_key(right_child);
-                            break true;
-                        }
-                        // No child on either side of the key can spare a key at this point, so
-                        // we merge them into one child, moving the
-                        // key into the merged child and try
-                        // to remove from this.
-                        self.merge(&mut node, index, &mut left_child, right_child);
-                        node = left_child;
-                        continue;
-                    }
-                    Err(index) => {
-                        // Node did not contain the key.
-                        if node.is_leaf() {
-                            break false;
-                        }
-                        // Node did not contain the key and is not a leaf.
-                        // Check and proactively prepare the child to be able to delete a key.
-                        node = self.prepare_child_for_key_removal(node, index);
-                    }
-                };
-            }
-        };
-
-        // If something was deleted, we update the length and make sure to remove the
-        // root node if needed.
-        if deleted_something {
-            self.len -= 1;
-            let root = self.get_node_mut(root_node_id);
-            if self.len == 0 {
-                // Remote the root node if tree is empty.
-                self.root = None;
-                self.delete_node(root);
-            } else {
-                // If the root is empty but the tree is not, point to the only child of the root
-                // as the root.
-                if root.keys.is_empty() {
-                    self.root = Some(root.children[0]);
-                    self.delete_node(root);
-                }
-            }
-        }
-        deleted_something
-    }
-
-    /// Internal function for taking the largest key in a subtree.
-    ///
-    /// Assumes:
-    /// - The provided node is not the root.
-    /// - The node contain more than minimum number of keys.
-    fn remove_largest_key<'a>(
-        &mut self,
-        mut node: StateRefMut<'a, state_btree_internals::Node<M, K>, S>,
-    ) -> K
-    where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        while !node.is_leaf() {
-            // Node is not a leaf, so we move further down this subtree.
-            let child_index = node.children.len() - 1;
-            node = self.prepare_child_for_key_removal(node, child_index);
-        }
-        // The node is a leaf, meaning we simply remove it.
-        // This will not violate the minimum keys invariant, since a node
-        // ensures a child can spare a key before iteration.
-        node.keys.pop().unwrap_abort()
-    }
-
-    /// Internal function for taking the smallest key in a subtree.
-    ///
-    /// Assumes:
-    /// - The provided node is not the root.
-    /// - The provided `node` contain more than minimum number of keys.
-    fn remove_smallest_key<'a>(
-        &mut self,
-        mut node: StateRefMut<'a, state_btree_internals::Node<M, K>, S>,
-    ) -> K
-    where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        while !node.is_leaf() {
-            // Node is not a leaf, so we move further down this subtree.
-            let child_index = 0;
-            node = self.prepare_child_for_key_removal(node, child_index);
-        }
-        // The node is a leaf, meaning we simply remove it.
-        // This will not violate the minimum keys invariant, since a node
-        // ensures a child can spare a key before iteration.
-        node.keys.remove(0)
-    }
-
-    /// Internal function for key rotation, preparing a node for deleting a key.
-    /// Returns the now prepared child at `index`.
-    /// Assumes:
-    /// - The provided `node` is not a leaf and has a child at `index`.
-    /// - The minimum degree `M` is at least 2 or more.
-    fn prepare_child_for_key_removal<'b, 'c>(
-        &mut self,
-        mut node: StateRefMut<'b, state_btree_internals::Node<M, K>, S>,
-        index: usize,
-    ) -> StateRefMut<'c, state_btree_internals::Node<M, K>, S>
-    where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        let mut child = self.get_node_mut(node.children[index]);
-        if !child.is_at_min() {
-            return child;
-        }
-        // The child is at minimum keys, so first attempt to take a key from either
-        // sibling, otherwise merge with one of them.
-        let has_smaller_sibling = 0 < index;
-        let has_larger_sibling = index < node.children.len() - 1;
-        let smaller_sibling = if has_smaller_sibling {
-            let mut smaller_sibling = self.get_node_mut(node.children[index - 1]);
-            if !smaller_sibling.is_at_min() {
-                // The smaller sibling can spare a key, so we replace the largest
-                // key from the sibling, put it in
-                // the parent and take a key from
-                // the parent.
-                let largest_key_sibling = smaller_sibling.keys.pop().unwrap_abort();
-                let swapped_node_key = mem::replace(&mut node.keys[index - 1], largest_key_sibling);
-                child.keys.insert(0, swapped_node_key);
-                if !child.is_leaf() {
-                    child.children.insert(0, smaller_sibling.children.pop().unwrap_abort());
-                }
-                return child;
-            }
-            Some(smaller_sibling)
-        } else {
-            None
-        };
-        let larger_sibling = if has_larger_sibling {
-            let mut larger_sibling = self.get_node_mut(node.children[index + 1]);
-            if !larger_sibling.is_at_min() {
-                // The larger sibling can spare a key, so we replace the smallest
-                // key from the sibling, put it in
-                // the parent and take a key from
-                // the parent.
-                let first_key_sibling = larger_sibling.keys.remove(0);
-                let swapped_node_key = mem::replace(&mut node.keys[index], first_key_sibling);
-                child.keys.push(swapped_node_key);
-
-                if !child.is_leaf() {
-                    child.children.push(larger_sibling.children.remove(0));
-                }
-                return child;
-            }
-            Some(larger_sibling)
-        } else {
-            None
-        };
-
-        if let Some(sibling) = larger_sibling {
-            self.merge(&mut node, index, &mut child, sibling);
-            child
-        } else if let Some(mut sibling) = smaller_sibling {
-            self.merge(&mut node, index - 1, &mut sibling, child);
-            sibling
-        } else {
-            // Unreachable code, since M must be 2 or larger (the minimum degree), a
-            // child node must have at least one sibling.
-            crate::trap();
-        }
-    }
-
-    /// Internal function for getting the highest key in a subtree.
-    /// Assumes the provided `node` is not a leaf and has a child at
-    /// `child_index`.
-    fn get_highest_key(&self, node: &state_btree_internals::Node<M, K>, child_index: usize) -> K
-    where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        let mut node = self.get_node(node.children[child_index]);
-        while !node.is_leaf() {
-            let child_node_id = node.children.last().unwrap_abort();
-            node = self.get_node(*child_node_id);
-        }
-        // This does not mutate the node in the end, just the representation in memory
-        // which is freed after the call.
-        node.keys.pop().unwrap_abort()
-    }
-
-    /// Internal function for getting the lowest key in a subtree.
-    /// Assumes the provided `node` is not a leaf and has a child at
-    /// `child_index`.
-    fn get_lowest_key(&self, node: &state_btree_internals::Node<M, K>, child_index: usize) -> K
-    where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        let mut node = self.get_node(node.children[child_index]);
-        while !node.is_leaf() {
-            let child_node_id = node.children.first().unwrap_abort();
-            node = self.get_node(*child_node_id);
-        }
-        // This does not mutate the node in the end, just the representation in memory
-        // which is freed after the call.
-        node.keys.swap_remove(0)
-    }
-
-    /// Moving key at `index` from the node to the lower child and then merges
-    /// this child with the content of its larger sibling, deleting the sibling.
-    ///
-    /// Assumes:
-    /// - `parent_node` has children `child` at `index` and `larger_child` at
-    ///   `index + 1`.
-    /// - Both children are at minimum number of keys (`M - 1`).
-    fn merge(
-        &mut self,
-        parent_node: &mut state_btree_internals::Node<M, K>,
-        index: usize,
-        child: &mut state_btree_internals::Node<M, K>,
-        mut larger_child: StateRefMut<state_btree_internals::Node<M, K>, S>,
-    ) where
-        K: Ord + Serialize,
-        S: HasStateApi, {
-        let parent_key = parent_node.keys.remove(index);
-        parent_node.children.remove(index + 1);
-        child.keys.push(parent_key);
-        child.keys.append(&mut larger_child.keys);
-        child.children.append(&mut larger_child.children);
-        self.delete_node(larger_child);
-    }
-
-    /// Internal function for constructing a node. It will increment the next
-    /// node ID and create an entry in the smart contract key-value store.
-    fn create_node<'b>(
-        &mut self,
-        keys: Vec<K>,
-        children: Vec<state_btree_internals::NodeId>,
-    ) -> (state_btree_internals::NodeId, StateRefMut<'b, state_btree_internals::Node<M, K>, S>)
-    where
-        K: Serialize,
-        S: HasStateApi, {
-        let node_id = self.next_node_id.copy_then_increment();
-        let node = state_btree_internals::Node {
-            keys,
-            children,
-        };
-        let entry = self.state_api.create_entry(&node_id.as_key(&self.prefix)).unwrap_abort();
-        let mut ref_mut: StateRefMut<'_, state_btree_internals::Node<M, K>, S> =
-            StateRefMut::new(entry, self.state_api.clone());
-        ref_mut.set(node);
-        (node_id, ref_mut)
-    }
-
-    /// Internal function for deleting a node, removing the entry in the smart
-    /// contract key-value store. Traps if no node was present.
-    fn delete_node(&mut self, node: StateRefMut<state_btree_internals::Node<M, K>, S>)
-    where
-        K: Serial,
-        S: HasStateApi, {
-        self.state_api.delete_entry(node.into_raw_parts().1).unwrap_abort()
-    }
-
-    /// Internal function for inserting into a subtree.
-    /// Assumes the given node is not full.
-    fn insert_non_full(
-        &mut self,
-        initial_node: StateRefMut<state_btree_internals::Node<M, K>, S>,
-        key: K,
-    ) -> bool
-    where
-        K: Serialize + Ord,
-        S: HasStateApi, {
-        let mut node = initial_node;
-        loop {
-            let Err(insert_index) = node.keys.binary_search(&key) else {
-                // We find the key in this node, so we do nothing.
-                return false;
-            };
-            // The key is not in this node.
-            if node.is_leaf() {
-                // Since we can assume the node is not full and this is a leaf, we can just
-                // insert here.
-                node.keys.insert(insert_index, key);
-                return true;
-            }
-
-            // The node is not a leaf, so we want to insert in the relevant child node.
-            let mut child = self.get_node_mut(node.children[insert_index]);
-            node = if !child.is_full() {
-                child
-            } else {
-                let larger_child = self.split_child(&mut node, insert_index, &mut child);
-                // Since the child is now split into two, we have to check which one to insert
-                // into.
-                if node.keys[insert_index] < key {
-                    larger_child
-                } else {
-                    child
-                }
-            };
-        }
-    }
-
-    /// Internal function for splitting the child node at a given index for a
-    /// given node. This will also mutate the given node adding a new key
-    /// and child after the provided child_index.
-    /// Returns the newly created node.
-    ///
-    /// Assumes:
-    /// - Node is not a leaf and has `child` as child at `child_index`.
-    /// - `child` is at maximum keys (`2 * M - 1`).
-    fn split_child<'b>(
-        &mut self,
-        node: &mut state_btree_internals::Node<M, K>,
-        child_index: usize,
-        child: &mut state_btree_internals::Node<M, K>,
-    ) -> StateRefMut<'b, state_btree_internals::Node<M, K>, S>
-    where
-        K: Serialize + Ord,
-        S: HasStateApi, {
-        let split_index = state_btree_internals::Node::<M, K>::MINIMUM_KEY_LEN + 1;
-        let (new_larger_sibling_id, new_larger_sibling) = self.create_node(
-            child.keys.split_off(split_index),
-            if child.is_leaf() {
-                Vec::new()
-            } else {
-                child.children.split_off(split_index)
-            },
-        );
-        let key = child.keys.pop().unwrap_abort();
-        node.children.insert(child_index + 1, new_larger_sibling_id);
-        node.keys.insert(child_index, key);
-        new_larger_sibling
-    }
-
-    /// Internal function for looking up a node in the tree.
-    /// This assumes the node is present and traps if this is not the case.
-    fn get_node<Key>(
-        &self,
-        node_id: state_btree_internals::NodeId,
-    ) -> state_btree_internals::Node<M, Key>
-    where
-        Key: Deserial,
-        S: HasStateApi, {
-        let key = node_id.as_key(&self.prefix);
-        let mut entry = self.state_api.lookup_entry(&key).unwrap_abort();
-        entry.get().unwrap_abort()
-    }
-
-    /// Internal function for looking up a node, providing mutable access.
-    /// This assumes the node is present and traps if this is not the case.
-    fn get_node_mut<'b>(
-        &mut self,
-        node_id: state_btree_internals::NodeId,
-    ) -> StateRefMut<'b, state_btree_internals::Node<M, K>, S>
-    where
-        K: Serial,
-        S: HasStateApi, {
-        let key = node_id.as_key(&self.prefix);
-        let entry = self.state_api.lookup_entry(&key).unwrap_abort();
-        StateRefMut::new(entry, self.state_api.clone())
-    }
-}
-
-/// Byte size of the key used to store a BTree internal node in the smart
-/// contract key-value store.
-const BTREE_NODE_KEY_SIZE: usize =
-    STATE_ITEM_PREFIX_SIZE + state_btree_internals::NodeId::SERIALIZED_BYTE_SIZE;
-
-impl<const M: usize, K> state_btree_internals::Node<M, K> {
-    /// The max length of the child list.
-    const MAXIMUM_CHILD_LEN: usize = 2 * M;
-    /// The max length of the key list.
-    const MAXIMUM_KEY_LEN: usize = Self::MAXIMUM_CHILD_LEN - 1;
-    /// The min length of the child list, when the node is not a leaf node.
-    const MINIMUM_CHILD_LEN: usize = M;
-    /// The min length of the key list, except when the node is root.
-    const MINIMUM_KEY_LEN: usize = Self::MINIMUM_CHILD_LEN - 1;
-
-    /// The number of keys stored in this node.
-    fn len(&self) -> usize { self.keys.len() }
-
-    /// Check if the node holds the maximum number of keys.
-    fn is_full(&self) -> bool { self.len() == Self::MAXIMUM_KEY_LEN }
-
-    /// Check if the node is representing a leaf in the tree.
-    fn is_leaf(&self) -> bool { self.children.is_empty() }
-
-    /// Check if the node holds the minimum number of keys.
-    fn is_at_min(&self) -> bool { self.len() == Self::MINIMUM_KEY_LEN }
-}
-
-impl state_btree_internals::NodeId {
-    /// Byte size of `NodeId` when serialized.
-    pub(crate) const SERIALIZED_BYTE_SIZE: usize = 4;
-
-    /// Return a copy of the NodeId, then increments itself.
-    pub(crate) fn copy_then_increment(&mut self) -> Self {
-        let current = *self;
-        self.id += 1;
-        current
-    }
-
-    /// Construct the key for the node in the key-value store from the node ID.
-    fn as_key(&self, prefix: &StateItemPrefix) -> [u8; BTREE_NODE_KEY_SIZE] {
-        // Create an uninitialized array of `MaybeUninit`. The `assume_init` is
-        // safe because the type we are claiming to have initialized here is a
-        // bunch of `MaybeUninit`s, which do not require initialization.
-        let mut prefixed: [MaybeUninit<u8>; BTREE_NODE_KEY_SIZE] =
-            unsafe { MaybeUninit::uninit().assume_init() };
-        for i in 0..STATE_ITEM_PREFIX_SIZE {
-            prefixed[i].write(prefix[i]);
-        }
-        let id_bytes = self.id.to_le_bytes();
-        for i in 0..id_bytes.len() {
-            prefixed[STATE_ITEM_PREFIX_SIZE + i].write(id_bytes[i]);
-        }
-        // Transmuting away the maybeuninit is safe since we have initialized all of
-        // them.
-        unsafe { mem::transmute(prefixed) }
-    }
-}
-
-impl<K: Deserial> Deserial for state_btree_internals::KeyWrapper<K> {
-    fn deserial<R: Read>(source: &mut R) -> ParseResult<Self> {
-        let key = K::deserial(source)?;
-        Ok(Self {
-            key: Some(key),
-        })
-    }
-}
-
-impl<'a, 'b, const M: usize, K, S> Iterator for StateBTreeSetIter<'a, 'b, K, S, M>
-where
-    'a: 'b,
-    K: Deserial,
-    S: HasStateApi,
-{
-    type Item = StateRef<'b, K>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        while let Some(id) = self.next_node.take() {
-            let node = self.tree.get_node(id);
-            if !node.is_leaf() {
-                self.next_node = Some(node.children[0]);
-            }
-            self.depth_first_stack.push((node, 0));
-        }
-
-        if let Some((node, index)) = self.depth_first_stack.last_mut() {
-            let key = node.keys[*index].key.take().unwrap_abort();
-            *index += 1;
-            let no_more_keys = index == &node.keys.len();
-            if !node.is_leaf() {
-                let child_id = node.children[*index];
-                self.next_node = Some(child_id);
-            }
-            if no_more_keys {
-                self.depth_first_stack.pop();
-            }
-            self.length -= 1;
-            Some(StateRef::new(key))
-        } else {
-            None
-        }
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) { (self.length, Some(self.length)) }
-}
-
-impl<'a, 'b, const M: usize, K, V, S> Iterator for StateBTreeMapIter<'a, 'b, K, V, S, M>
-where
-    'a: 'b,
-    K: Serialize,
-    V: Serial + DeserialWithState<S> + 'b,
-    S: HasStateApi,
-{
-    type Item = (StateRef<'b, K>, StateRef<'b, V>);
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let next_key = self.key_iter.next()?;
-        let value = self.map.get(&next_key).unwrap_abort();
-        // Unwrap is safe, otherwise the map and the set have inconsistencies.
-        Some((next_key, value))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) { self.key_iter.size_hint() }
-}
-
-impl<const M: usize, K, V, S> Deletable for StateBTreeMap<K, V, S, M>
-where
-    S: HasStateApi,
-    K: Serialize,
-    V: Serial + DeserialWithState<S> + Deletable,
-{
-    fn delete(mut self) { self.clear(); }
-}
-
 #[cfg(test)]
 mod tests {
 
@@ -4196,19 +3153,469 @@ mod tests {
     }
 }
 
-#[concordium_cfg_test]
+/// This test module rely on the runtime providing host functions and can only
+/// be run using `cargo concordium test`.
+#[cfg(feature = "wasm-test")]
 mod wasm_test {
-    use concordium_contracts_common::concordium_test;
+    use crate::{
+        claim, claim_eq, concordium_test, to_bytes, Deletable, Deserial, DeserialWithState,
+        EntryRaw, HasStateApi, HasStateEntry, ParseResult, Serial, StateApi, StateBuilder,
+        StateError, StateMap, StateSet, INITIAL_NEXT_ITEM_PREFIX,
+    };
 
-    use crate::{claim_eq, StateApi, StateBuilder};
+    const GENERIC_MAP_PREFIX: u64 = 1;
+
+    /// Some helper methods that are used for internal tests.
+    impl<S> StateBuilder<S>
+    where
+        S: HasStateApi,
+    {
+        /// Get a value from the generic map.
+        /// `Some(Err(_))` means that something exists in the state with that
+        /// key, but it isn't of type `V`.
+        pub(crate) fn get<K: Serial, V: DeserialWithState<S>>(
+            &self,
+            key: K,
+        ) -> Option<ParseResult<V>> {
+            let key_with_map_prefix = Self::prepend_generic_map_key(key);
+
+            self.state_api
+                .lookup_entry(&key_with_map_prefix)
+                .map(|mut entry| V::deserial_with_state(&self.state_api, &mut entry))
+        }
+
+        /// Inserts a value in the generic map.
+        /// The key and value are serialized before insert.
+        pub(crate) fn insert<K: Serial, V: Serial>(
+            &mut self,
+            key: K,
+            value: V,
+        ) -> Result<(), StateError> {
+            let key_with_map_prefix = Self::prepend_generic_map_key(key);
+            match self.state_api.entry(key_with_map_prefix) {
+                EntryRaw::Vacant(vac) => {
+                    let _ = vac.insert(&value);
+                }
+                EntryRaw::Occupied(mut occ) => occ.insert(&value),
+            }
+            Ok(())
+        }
+
+        /// Serializes the key and prepends [GENERIC_MAP_PREFIX].
+        /// This is similar to how [StateMap] works, where a unique prefix is
+        /// prepended onto keys. Since there is only one generic map, the prefix
+        /// is a constant.
+        fn prepend_generic_map_key<K: Serial>(key: K) -> Vec<u8> {
+            let mut key_with_map_prefix = to_bytes(&GENERIC_MAP_PREFIX);
+            key_with_map_prefix.append(&mut to_bytes(&key));
+            key_with_map_prefix
+        }
+    }
 
     #[concordium_test]
-    fn test() {
+    fn high_level_insert_get() {
+        let expected_value: u64 = 123123123;
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        state_builder.insert(0, expected_value).expect("Insert failed");
+        let actual_value: u64 = state_builder.get(0).expect("Not found").expect("Not a valid u64");
+        claim_eq!(expected_value, actual_value);
+    }
+
+    #[concordium_test]
+    fn low_level_entry() {
+        let expected_value: u64 = 123123123;
+        let key = to_bytes(&42u64);
+        let mut state = StateApi::open();
+        state
+            .entry(&key[..])
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("No iterators, so insertion should work.");
+
+        match state.entry(key) {
+            EntryRaw::Vacant(_) => panic!("Unexpected vacant entry."),
+            EntryRaw::Occupied(occ) => {
+                claim_eq!(u64::deserial(&mut occ.get()), Ok(expected_value))
+            }
+        }
+    }
+
+    #[concordium_test]
+    fn high_level_statemap() {
+        let my_map_key = "my_map";
+        let mut state_builder = StateBuilder::open(StateApi::open());
+
+        let map_to_insert = state_builder.new_map::<String, String>();
+        state_builder.insert(my_map_key, map_to_insert).expect("Insert failed");
+
+        let mut my_map: StateMap<String, String, _> = state_builder
+            .get(my_map_key)
+            .expect("Could not get statemap")
+            .expect("Deserializing statemap failed");
+        my_map.insert("abc".to_string(), "hello, world".to_string());
+        my_map.insert("def".to_string(), "hallo, Weld".to_string());
+        my_map.insert("ghi".to_string(), "hej, verden".to_string());
+        claim_eq!(*my_map.get(&"abc".to_string()).unwrap(), "hello, world".to_string());
+
+        let mut iter = my_map.iter();
+        let (k1, v1) = iter.next().unwrap();
+        claim_eq!(*k1, "abc".to_string());
+        claim_eq!(*v1, "hello, world".to_string());
+        let (k2, v2) = iter.next().unwrap();
+        claim_eq!(*k2, "def".to_string());
+        claim_eq!(*v2, "hallo, Weld".to_string());
+        let (k3, v3) = iter.next().unwrap();
+        claim_eq!(*k3, "ghi".to_string());
+        claim_eq!(*v3, "hej, verden".to_string());
+        claim!(iter.next().is_none());
+    }
+
+    #[concordium_test]
+    fn statemap_insert_remove() {
         let mut state_builder = StateBuilder::open(StateApi::open());
         let mut map = state_builder.new_map();
+        let value = String::from("hello");
+        let _ = map.insert(42, value.clone());
+        claim_eq!(*map.get(&42).unwrap(), value);
+        map.remove(&42);
+        claim!(map.get(&42).is_none());
+    }
 
-        map.insert(0, "0".to_string());
+    #[concordium_test]
+    fn statemap_clear() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut map = state_builder.new_map();
+        let _ = map.insert(1, 2);
+        let _ = map.insert(2, 3);
+        let _ = map.insert(3, 4);
+        map.clear();
+        claim!(map.is_empty());
+    }
 
-        claim_eq!(map.get(&0).as_deref(), Some(&"1".to_string()));
+    #[concordium_test]
+    fn high_level_nested_statemaps() {
+        let inner_map_key = 0u8;
+        let key_to_value = 77u8;
+        let value = 255u8;
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut outer_map = state_builder.new_map::<u8, StateMap<u8, u8, _>>();
+        let mut inner_map = state_builder.new_map::<u8, u8>();
+
+        inner_map.insert(key_to_value, value);
+        outer_map.insert(inner_map_key, inner_map);
+
+        claim_eq!(*outer_map.get(&inner_map_key).unwrap().get(&key_to_value).unwrap(), value);
+    }
+
+    #[concordium_test]
+    fn statemap_iter_mut_works() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut map = state_builder.new_map();
+        map.insert(0u8, 1u8);
+        map.insert(1u8, 2u8);
+        map.insert(2u8, 3u8);
+        for (_, mut v) in map.iter_mut() {
+            v.update(|old_value| *old_value += 10);
+        }
+        let mut iter = map.iter();
+        let (k1, v1) = iter.next().unwrap();
+        claim_eq!(*k1, 0);
+        claim_eq!(*v1, 11);
+        let (k2, v2) = iter.next().unwrap();
+        claim_eq!(*k2, 1);
+        claim_eq!(*v2, 12);
+        let (k3, v3) = iter.next().unwrap();
+        claim_eq!(*k3, 2);
+        claim_eq!(*v3, 13);
+        claim!(iter.next().is_none());
+    }
+
+    #[concordium_test]
+    fn iter_mut_works_on_nested_statemaps() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut outer_map = state_builder.new_map();
+        let mut inner_map = state_builder.new_map();
+        inner_map.insert(0u8, 1u8);
+        inner_map.insert(1u8, 2u8);
+        outer_map.insert(99u8, inner_map);
+        for (_, mut v_map) in outer_map.iter_mut() {
+            v_map.update(|v_map| {
+                for (_, mut inner_v) in v_map.iter_mut() {
+                    inner_v.update(|inner_v| *inner_v += 10);
+                }
+            });
+        }
+
+        // Check the outer map.
+        let mut outer_iter = outer_map.iter();
+        let (inner_map_key, inner_map) = outer_iter.next().unwrap();
+        claim_eq!(*inner_map_key, 99);
+        claim!(outer_iter.next().is_none());
+
+        // Check the inner map.
+        let mut inner_iter = inner_map.iter();
+        let (k1, v1) = inner_iter.next().unwrap();
+        claim_eq!(*k1, 0);
+        claim_eq!(*v1, 11);
+        let (k2, v2) = inner_iter.next().unwrap();
+        claim_eq!(*k2, 1);
+        claim_eq!(*v2, 12);
+        claim!(inner_iter.next().is_none());
+    }
+
+    #[concordium_test]
+    fn statemap_iterator_unlocks_tree_once_dropped() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut map = state_builder.new_map();
+        map.insert(0u8, 1u8);
+        map.insert(1u8, 2u8);
+        {
+            let _iter = map.iter();
+            // Uncommenting these two lines (and making iter mutable) should
+            // give a compile error:
+            //
+            // map.insert(2u8, 3u8);
+            // let n = iter.next();
+        } // iter is dropped here, unlocking the subtree.
+        map.insert(2u8, 3u8);
+    }
+
+    #[concordium_test]
+    fn high_level_stateset() {
+        let my_set_key = "my_set";
+        let mut state_builder = StateBuilder::open(StateApi::open());
+
+        let mut set = state_builder.new_set::<u8>();
+        claim!(set.insert(0));
+        claim!(set.insert(1));
+        claim!(!set.insert(1));
+        claim!(set.insert(2));
+        claim!(set.remove(&2));
+        state_builder.insert(my_set_key, set).expect("Insert failed");
+
+        claim!(state_builder.get::<_, StateSet<u8, _>>(my_set_key).unwrap().unwrap().contains(&0),);
+        claim!(!state_builder.get::<_, StateSet<u8, _>>(my_set_key).unwrap().unwrap().contains(&2),);
+
+        let set = state_builder.get::<_, StateSet<u8, _>>(my_set_key).unwrap().unwrap();
+        let mut iter = set.iter();
+        claim_eq!(*iter.next().unwrap(), 0);
+        claim_eq!(*iter.next().unwrap(), 1);
+        claim!(iter.next().is_none());
+    }
+
+    #[concordium_test]
+    fn high_level_nested_stateset() {
+        let inner_set_key = 0u8;
+        let value = 255u8;
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut outer_map = state_builder.new_map::<u8, StateSet<u8, _>>();
+        let mut inner_set = state_builder.new_set::<u8>();
+
+        inner_set.insert(value);
+        outer_map.insert(inner_set_key, inner_set);
+
+        claim!(outer_map.get(&inner_set_key).unwrap().contains(&value));
+    }
+
+    #[concordium_test]
+    fn stateset_insert_remove() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut set = state_builder.new_set();
+        let _ = set.insert(42);
+        claim!(set.contains(&42));
+        set.remove(&42);
+        claim!(!set.contains(&42));
+    }
+
+    #[concordium_test]
+    fn stateset_clear() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut set = state_builder.new_set();
+        let _ = set.insert(1);
+        let _ = set.insert(2);
+        let _ = set.insert(3);
+        set.clear();
+        claim!(set.is_empty());
+    }
+
+    #[concordium_test]
+    fn stateset_iterator_unlocks_tree_once_dropped() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut set = state_builder.new_set();
+        set.insert(0u8);
+        set.insert(1);
+        {
+            let _iter = set.iter();
+            // Uncommenting these two lines (and making iter mutable) should
+            // give a compile error:
+            //
+            // set.insert(2);
+            // let n = iter.next();
+        } // iter is dropped here, unlocking the subtree.
+        set.insert(2);
+    }
+
+    #[concordium_test]
+    fn allocate_and_get_statebox() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let boxed_value = String::from("I'm boxed");
+        let statebox = state_builder.new_box(boxed_value.clone());
+        claim_eq!(*statebox.get(), boxed_value);
+    }
+
+    #[concordium_test]
+    fn a_new_entry_can_not_be_created_under_a_locked_subtree() {
+        let expected_value: u64 = 123123123;
+        let key = to_bytes(b"ab");
+        let sub_key = to_bytes(b"abc");
+        let mut state = StateApi::open();
+        state
+            .entry(&key[..])
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("No iterators, so insertion should work.");
+        claim!(state.iterator(&key).is_ok(), "Iterator should be present");
+        let entry = state.create_entry(&sub_key);
+        claim!(entry.is_err(), "Should not be able to create an entry under a locked subtree");
+    }
+
+    #[concordium_test]
+    fn a_new_entry_can_be_created_under_a_different_subtree_in_same_super_tree() {
+        let expected_value: u64 = 123123123;
+        let key = to_bytes(b"abcd");
+        let key2 = to_bytes(b"abe");
+        let mut state = StateApi::open();
+        state
+            .entry(&key[..])
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("No iterators, so insertion should work.");
+        claim!(state.iterator(&key).is_ok(), "Iterator should be present");
+        let entry = state.create_entry(&key2);
+        claim!(entry.is_ok(), "Failed to create a new entry under a different subtree");
+    }
+
+    #[concordium_test]
+    fn an_existing_entry_can_not_be_deleted_under_a_locked_subtree() {
+        let expected_value: u64 = 123123123;
+        let key = to_bytes(b"ab");
+        let sub_key = to_bytes(b"abc");
+        let mut state = StateApi::open();
+        state
+            .entry(&key[..])
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("no iterators, so insertion should work.");
+        let sub_entry = state
+            .entry(sub_key)
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("Should be possible to create the entry.");
+        claim!(state.iterator(&key).is_ok(), "Iterator should be present");
+        claim!(
+            state.delete_entry(sub_entry).is_err(),
+            "Should not be able to create an entry under a locked subtree"
+        );
+    }
+
+    #[concordium_test]
+    fn an_existing_entry_can_be_deleted_from_a_different_subtree_in_same_super_tree() {
+        let expected_value: u64 = 123123123;
+        let key = to_bytes(b"abcd");
+        let key2 = to_bytes(b"abe");
+        let mut state = StateApi::open();
+        state
+            .entry(&key[..])
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("No iterators, so insertion should work.");
+        let entry2 = state
+            .entry(key2)
+            .or_insert_raw(&to_bytes(&expected_value))
+            .expect("Should be possible to create the entry.");
+        claim!(state.iterator(&key).is_ok(), "Iterator should be present");
+        claim!(
+            state.delete_entry(entry2).is_ok(),
+            "Failed to create a new entry under a different subtree"
+        );
+    }
+
+    #[concordium_test]
+    fn deleting_nested_stateboxes_works() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let inner_box = state_builder.new_box(99u8);
+        let middle_box = state_builder.new_box(inner_box);
+        let outer_box = state_builder.new_box(middle_box);
+        outer_box.delete();
+        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
+        // The only remaining node should be the state_builder's next_item_prefix node.
+        claim!(iter.nth(1).is_none());
+    }
+
+    #[concordium_test]
+    fn clearing_statemap_with_stateboxes_works() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let box1 = state_builder.new_box(1u8);
+        let box2 = state_builder.new_box(2u8);
+        let box3 = state_builder.new_box(3u8);
+        let mut map = state_builder.new_map();
+        map.insert(1u8, box1);
+        map.insert(2u8, box2);
+        map.insert(3u8, box3);
+        map.clear();
+        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
+        // The only remaining node should be the state_builder's next_item_prefix node.
+        claim!(iter.nth(1).is_none());
+    }
+
+    #[concordium_test]
+    fn clearing_nested_statemaps_works() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut inner_map_1 = state_builder.new_map();
+        inner_map_1.insert(1u8, 2u8);
+        inner_map_1.insert(2u8, 3u8);
+        inner_map_1.insert(3u8, 4u8);
+        let mut inner_map_2 = state_builder.new_map();
+        inner_map_2.insert(11u8, 12u8);
+        inner_map_2.insert(12u8, 13u8);
+        inner_map_2.insert(13u8, 14u8);
+        let mut outer_map = state_builder.new_map();
+        outer_map.insert(0u8, inner_map_1);
+        outer_map.insert(1u8, inner_map_2);
+        outer_map.clear();
+        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
+        // The only remaining node should be the state_builder's next_item_prefix node.
+        claim!(iter.nth(1).is_none());
+    }
+
+    #[concordium_test]
+    fn occupied_entry_truncates_leftover_data() {
+        let mut state_builder = StateBuilder::open(StateApi::open());
+        let mut map = state_builder.new_map();
+        map.insert(99u8, "A longer string that should be truncated".into());
+        let a_short_string = "A short string".to_string();
+        let expected_size = a_short_string.len() + 4; // 4 bytes for the length of the string.
+        map.entry(99u8).and_modify(|v| *v = a_short_string);
+        let actual_size = state_builder
+            .state_api
+            .lookup_entry(&[INITIAL_NEXT_ITEM_PREFIX[0], 0, 0, 0, 0, 0, 0, 0, 99])
+            .expect("Lookup failed")
+            .size()
+            .expect("Getting size failed");
+        claim_eq!(expected_size as u32, actual_size);
+    }
+
+    #[concordium_test]
+    fn occupied_entry_raw_truncates_leftover_data() {
+        let mut state = StateApi::open();
+        state
+            .entry([])
+            .or_insert_raw(&to_bytes(&"A longer string that should be truncated"))
+            .expect("No iterators, so insertion should work.");
+
+        let a_short_string = "A short string";
+        let expected_size = a_short_string.len() + 4; // 4 bytes for the length of the string.
+
+        match state.entry([]) {
+            EntryRaw::Vacant(_) => panic!("Entry is vacant"),
+            EntryRaw::Occupied(mut occ) => occ.insert_raw(&to_bytes(&a_short_string)),
+        }
+        let actual_size =
+            state.lookup_entry(&[]).expect("Lookup failed").size().expect("Getting size failed");
+        claim_eq!(expected_size as u32, actual_size);
     }
 }
