+++ concordium-std/src/test_infrastructure.rs
@@ -2076,640 +2076,4 @@ mod test {
             "Correct data was written."
         );
     }
-
-    #[test]
-    fn high_level_insert_get() {
-        let expected_value: u64 = 123123123;
-        let mut state_builder = TestStateBuilder::new();
-        state_builder.insert(0, expected_value).expect("Insert failed");
-        let actual_value: u64 = state_builder.get(0).expect("Not found").expect("Not a valid u64");
-        assert_eq!(expected_value, actual_value);
-    }
-
-    #[test]
-    fn low_level_entry() {
-        let expected_value: u64 = 123123123;
-        let key = to_bytes(&42u64);
-        let mut state = TestStateApi::new();
-        state
-            .entry(&key[..])
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("No iterators, so insertion should work.");
-
-        match state.entry(key) {
-            EntryRaw::Vacant(_) => panic!("Unexpected vacant entry."),
-            EntryRaw::Occupied(occ) => {
-                assert_eq!(u64::deserial(&mut occ.get()), Ok(expected_value))
-            }
-        }
-    }
-
-    #[test]
-    fn high_level_statemap() {
-        let my_map_key = "my_map";
-        let mut state_builder = TestStateBuilder::new();
-
-        let map_to_insert = state_builder.new_map::<String, String>();
-        state_builder.insert(my_map_key, map_to_insert).expect("Insert failed");
-
-        let mut my_map: StateMap<String, String, _> = state_builder
-            .get(my_map_key)
-            .expect("Could not get statemap")
-            .expect("Deserializing statemap failed");
-        my_map.insert("abc".to_string(), "hello, world".to_string());
-        my_map.insert("def".to_string(), "hallo, Weld".to_string());
-        my_map.insert("ghi".to_string(), "hej, verden".to_string());
-        assert_eq!(*my_map.get(&"abc".to_string()).unwrap(), "hello, world".to_string());
-
-        let mut iter = my_map.iter();
-        let (k1, v1) = iter.next().unwrap();
-        assert_eq!(*k1, "abc".to_string());
-        assert_eq!(*v1, "hello, world".to_string());
-        let (k2, v2) = iter.next().unwrap();
-        assert_eq!(*k2, "def".to_string());
-        assert_eq!(*v2, "hallo, Weld".to_string());
-        let (k3, v3) = iter.next().unwrap();
-        assert_eq!(*k3, "ghi".to_string());
-        assert_eq!(*v3, "hej, verden".to_string());
-        assert!(iter.next().is_none());
-    }
-
-    #[test]
-    fn statemap_insert_remove() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut map = state_builder.new_map();
-        let value = String::from("hello");
-        let _ = map.insert(42, value.clone());
-        assert_eq!(*map.get(&42).unwrap(), value);
-        map.remove(&42);
-        assert!(map.get(&42).is_none());
-    }
-
-    #[test]
-    fn statemap_clear() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut map = state_builder.new_map();
-        let _ = map.insert(1, 2);
-        let _ = map.insert(2, 3);
-        let _ = map.insert(3, 4);
-        map.clear();
-        assert!(map.is_empty());
-    }
-
-    #[test]
-    fn high_level_nested_statemaps() {
-        let inner_map_key = 0u8;
-        let key_to_value = 77u8;
-        let value = 255u8;
-        let mut state_builder = TestStateBuilder::new();
-        let mut outer_map = state_builder.new_map::<u8, StateMap<u8, u8, _>>();
-        let mut inner_map = state_builder.new_map::<u8, u8>();
-
-        inner_map.insert(key_to_value, value);
-        outer_map.insert(inner_map_key, inner_map);
-
-        assert_eq!(*outer_map.get(&inner_map_key).unwrap().get(&key_to_value).unwrap(), value);
-    }
-
-    #[test]
-    fn statemap_iter_mut_works() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut map = state_builder.new_map();
-        map.insert(0u8, 1u8);
-        map.insert(1u8, 2u8);
-        map.insert(2u8, 3u8);
-        for (_, mut v) in map.iter_mut() {
-            v.update(|old_value| *old_value += 10);
-        }
-        let mut iter = map.iter();
-        let (k1, v1) = iter.next().unwrap();
-        assert_eq!(*k1, 0);
-        assert_eq!(*v1, 11);
-        let (k2, v2) = iter.next().unwrap();
-        assert_eq!(*k2, 1);
-        assert_eq!(*v2, 12);
-        let (k3, v3) = iter.next().unwrap();
-        assert_eq!(*k3, 2);
-        assert_eq!(*v3, 13);
-        assert!(iter.next().is_none());
-    }
-
-    #[test]
-    fn iter_mut_works_on_nested_statemaps() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut outer_map = state_builder.new_map();
-        let mut inner_map = state_builder.new_map();
-        inner_map.insert(0u8, 1u8);
-        inner_map.insert(1u8, 2u8);
-        outer_map.insert(99u8, inner_map);
-        for (_, mut v_map) in outer_map.iter_mut() {
-            v_map.update(|v_map| {
-                for (_, mut inner_v) in v_map.iter_mut() {
-                    inner_v.update(|inner_v| *inner_v += 10);
-                }
-            });
-        }
-
-        // Check the outer map.
-        let mut outer_iter = outer_map.iter();
-        let (inner_map_key, inner_map) = outer_iter.next().unwrap();
-        assert_eq!(*inner_map_key, 99);
-        assert!(outer_iter.next().is_none());
-
-        // Check the inner map.
-        let mut inner_iter = inner_map.iter();
-        let (k1, v1) = inner_iter.next().unwrap();
-        assert_eq!(*k1, 0);
-        assert_eq!(*v1, 11);
-        let (k2, v2) = inner_iter.next().unwrap();
-        assert_eq!(*k2, 1);
-        assert_eq!(*v2, 12);
-        assert!(inner_iter.next().is_none());
-    }
-
-    #[test]
-    fn statemap_iterator_unlocks_tree_once_dropped() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut map = state_builder.new_map();
-        map.insert(0u8, 1u8);
-        map.insert(1u8, 2u8);
-        {
-            let _iter = map.iter();
-            // Uncommenting these two lines (and making iter mutable) should
-            // give a compile error:
-            //
-            // map.insert(2u8, 3u8);
-            // let n = iter.next();
-        } // iter is dropped here, unlocking the subtree.
-        map.insert(2u8, 3u8);
-    }
-
-    #[test]
-    fn high_level_stateset() {
-        let my_set_key = "my_set";
-        let mut state_builder = TestStateBuilder::new();
-
-        let mut set = state_builder.new_set::<u8>();
-        assert!(set.insert(0));
-        assert!(set.insert(1));
-        assert!(!set.insert(1));
-        assert!(set.insert(2));
-        assert!(set.remove(&2));
-        state_builder.insert(my_set_key, set).expect("Insert failed");
-
-        assert!(state_builder.get::<_, StateSet<u8, _>>(my_set_key).unwrap().unwrap().contains(&0),);
-        assert!(!state_builder
-            .get::<_, StateSet<u8, _>>(my_set_key)
-            .unwrap()
-            .unwrap()
-            .contains(&2),);
-
-        let set = state_builder.get::<_, StateSet<u8, _>>(my_set_key).unwrap().unwrap();
-        let mut iter = set.iter();
-        assert_eq!(*iter.next().unwrap(), 0);
-        assert_eq!(*iter.next().unwrap(), 1);
-        assert!(iter.next().is_none());
-    }
-
-    #[test]
-    fn high_level_nested_stateset() {
-        let inner_set_key = 0u8;
-        let value = 255u8;
-        let mut state_builder = TestStateBuilder::new();
-        let mut outer_map = state_builder.new_map::<u8, StateSet<u8, _>>();
-        let mut inner_set = state_builder.new_set::<u8>();
-
-        inner_set.insert(value);
-        outer_map.insert(inner_set_key, inner_set);
-
-        assert!(outer_map.get(&inner_set_key).unwrap().contains(&value));
-    }
-
-    #[test]
-    fn stateset_insert_remove() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut set = state_builder.new_set();
-        let _ = set.insert(42);
-        assert!(set.contains(&42));
-        set.remove(&42);
-        assert!(!set.contains(&42));
-    }
-
-    #[test]
-    fn stateset_clear() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut set = state_builder.new_set();
-        let _ = set.insert(1);
-        let _ = set.insert(2);
-        let _ = set.insert(3);
-        set.clear();
-        assert!(set.is_empty());
-    }
-
-    #[test]
-    fn stateset_iterator_unlocks_tree_once_dropped() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut set = state_builder.new_set();
-        set.insert(0u8);
-        set.insert(1);
-        {
-            let _iter = set.iter();
-            // Uncommenting these two lines (and making iter mutable) should
-            // give a compile error:
-            //
-            // set.insert(2);
-            // let n = iter.next();
-        } // iter is dropped here, unlocking the subtree.
-        set.insert(2);
-    }
-
-    #[test]
-    fn allocate_and_get_statebox() {
-        let mut state_builder = TestStateBuilder::new();
-        let boxed_value = String::from("I'm boxed");
-        let statebox = state_builder.new_box(boxed_value.clone());
-        assert_eq!(*statebox.get(), boxed_value);
-    }
-
-    #[test]
-    fn a_new_entry_can_not_be_created_under_a_locked_subtree() {
-        let expected_value: u64 = 123123123;
-        let key = to_bytes(b"ab");
-        let sub_key = to_bytes(b"abc");
-        let mut state = TestStateApi::new();
-        state
-            .entry(&key[..])
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("No iterators, so insertion should work.");
-        assert!(state.iterator(&key).is_ok(), "Iterator should be present");
-        let entry = state.create_entry(&sub_key);
-        assert!(entry.is_err(), "Should not be able to create an entry under a locked subtree");
-    }
-
-    #[test]
-    fn a_new_entry_can_be_created_under_a_different_subtree_in_same_super_tree() {
-        let expected_value: u64 = 123123123;
-        let key = to_bytes(b"abcd");
-        let key2 = to_bytes(b"abe");
-        let mut state = TestStateApi::new();
-        state
-            .entry(&key[..])
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("No iterators, so insertion should work.");
-        assert!(state.iterator(&key).is_ok(), "Iterator should be present");
-        let entry = state.create_entry(&key2);
-        assert!(entry.is_ok(), "Failed to create a new entry under a different subtree");
-    }
-
-    #[test]
-    fn an_existing_entry_can_not_be_deleted_under_a_locked_subtree() {
-        let expected_value: u64 = 123123123;
-        let key = to_bytes(b"ab");
-        let sub_key = to_bytes(b"abc");
-        let mut state = TestStateApi::new();
-        state
-            .entry(&key[..])
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("no iterators, so insertion should work.");
-        let sub_entry = state
-            .entry(sub_key)
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("Should be possible to create the entry.");
-        assert!(state.iterator(&key).is_ok(), "Iterator should be present");
-        assert!(
-            state.delete_entry(sub_entry).is_err(),
-            "Should not be able to create an entry under a locked subtree"
-        );
-    }
-
-    #[test]
-    fn an_existing_entry_can_be_deleted_from_a_different_subtree_in_same_super_tree() {
-        let expected_value: u64 = 123123123;
-        let key = to_bytes(b"abcd");
-        let key2 = to_bytes(b"abe");
-        let mut state = TestStateApi::new();
-        state
-            .entry(&key[..])
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("No iterators, so insertion should work.");
-        let entry2 = state
-            .entry(key2)
-            .or_insert_raw(&to_bytes(&expected_value))
-            .expect("Should be possible to create the entry.");
-        assert!(state.iterator(&key).is_ok(), "Iterator should be present");
-        assert!(
-            state.delete_entry(entry2).is_ok(),
-            "Failed to create a new entry under a different subtree"
-        );
-    }
-
-    #[test]
-    fn deleting_nested_stateboxes_works() {
-        let mut state_builder = TestStateBuilder::new();
-        let inner_box = state_builder.new_box(99u8);
-        let middle_box = state_builder.new_box(inner_box);
-        let outer_box = state_builder.new_box(middle_box);
-        outer_box.delete();
-        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
-        // The only remaining node should be the state_builder's next_item_prefix node.
-        assert!(iter.nth(1).is_none());
-    }
-
-    #[test]
-    fn clearing_statemap_with_stateboxes_works() {
-        let mut state_builder = TestStateBuilder::new();
-        let box1 = state_builder.new_box(1u8);
-        let box2 = state_builder.new_box(2u8);
-        let box3 = state_builder.new_box(3u8);
-        let mut map = state_builder.new_map();
-        map.insert(1u8, box1);
-        map.insert(2u8, box2);
-        map.insert(3u8, box3);
-        map.clear();
-        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
-        // The only remaining node should be the state_builder's next_item_prefix node.
-        assert!(iter.nth(1).is_none());
-    }
-
-    #[test]
-    fn clearing_nested_statemaps_works() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut inner_map_1 = state_builder.new_map();
-        inner_map_1.insert(1u8, 2u8);
-        inner_map_1.insert(2u8, 3u8);
-        inner_map_1.insert(3u8, 4u8);
-        let mut inner_map_2 = state_builder.new_map();
-        inner_map_2.insert(11u8, 12u8);
-        inner_map_2.insert(12u8, 13u8);
-        inner_map_2.insert(13u8, 14u8);
-        let mut outer_map = state_builder.new_map();
-        outer_map.insert(0u8, inner_map_1);
-        outer_map.insert(1u8, inner_map_2);
-        outer_map.clear();
-        let mut iter = state_builder.state_api.iterator(&[]).expect("Could not get iterator");
-        // The only remaining node should be the state_builder's next_item_prefix node.
-        assert!(iter.nth(1).is_none());
-    }
-
-    #[test]
-    fn occupied_entry_truncates_leftover_data() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut map = state_builder.new_map();
-        map.insert(99u8, "A longer string that should be truncated".into());
-        let a_short_string = "A short string".to_string();
-        let expected_size = a_short_string.len() + 4; // 4 bytes for the length of the string.
-        map.entry(99u8).and_modify(|v| *v = a_short_string);
-        let actual_size = state_builder
-            .state_api
-            .lookup_entry(&[INITIAL_NEXT_ITEM_PREFIX[0], 0, 0, 0, 0, 0, 0, 0, 99])
-            .expect("Lookup failed")
-            .size()
-            .expect("Getting size failed");
-        assert_eq!(expected_size as u32, actual_size);
-    }
-
-    #[test]
-    fn occupied_entry_raw_truncates_leftover_data() {
-        let mut state = TestStateApi::new();
-        state
-            .entry([])
-            .or_insert_raw(&to_bytes(&"A longer string that should be truncated"))
-            .expect("No iterators, so insertion should work.");
-
-        let a_short_string = "A short string";
-        let expected_size = a_short_string.len() + 4; // 4 bytes for the length of the string.
-
-        match state.entry([]) {
-            EntryRaw::Vacant(_) => panic!("Entry is vacant"),
-            EntryRaw::Occupied(mut occ) => occ.insert_raw(&to_bytes(&a_short_string)),
-        }
-        let actual_size =
-            state.lookup_entry(&[]).expect("Lookup failed").size().expect("Getting size failed");
-        assert_eq!(expected_size as u32, actual_size);
-    }
-
-    #[test]
-    fn test_btree_insert_6() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<5, _>();
-        for n in 0..=5 {
-            tree.insert(n);
-        }
-        for n in 0..=5 {
-            assert!(tree.contains(&n));
-        }
-    }
-
-    #[test]
-    fn test_btree_insert_0_7() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..=7 {
-            tree.insert(n);
-        }
-        for n in 0..=7 {
-            assert!(tree.contains(&n));
-        }
-    }
-
-    #[test]
-    fn test_btree_insert_7_no_order() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-
-        tree.insert(0);
-        tree.insert(1);
-        tree.insert(2);
-        tree.insert(3);
-        tree.insert(7);
-        tree.insert(6);
-        tree.insert(5);
-        tree.insert(4);
-
-        for n in 0..=7 {
-            assert!(tree.contains(&n));
-        }
-    }
-
-    #[test]
-    fn test_btree_higher() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        tree.insert(1);
-        tree.insert(2);
-        tree.insert(3);
-        tree.insert(4);
-        tree.insert(5);
-        tree.insert(7);
-        assert_eq!(tree.higher(&3).as_deref(), Some(&4));
-        assert_eq!(tree.higher(&5).as_deref(), Some(&7));
-        assert_eq!(tree.higher(&7).as_deref(), None)
-    }
-
-    #[test]
-    fn test_btree_lower() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        tree.insert(1);
-        tree.insert(2);
-        tree.insert(3);
-        tree.insert(4);
-        tree.insert(5);
-        tree.insert(7);
-        assert_eq!(tree.lower(&3).as_deref(), Some(&2));
-        assert_eq!(tree.lower(&7).as_deref(), Some(&5));
-        assert_eq!(tree.lower(&1).as_deref(), None)
-    }
-
-    #[test]
-    fn test_btree_insert_1000() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..500 {
-            tree.insert(n);
-        }
-
-        for n in (500..1000).into_iter().rev() {
-            tree.insert(n);
-        }
-
-        for n in 0..1000 {
-            assert!(tree.contains(&n))
-        }
-
-        assert_eq!(tree.len(), 1000)
-    }
-
-    #[test]
-    fn test_btree_7_get_8() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..=7 {
-            tree.insert(n);
-        }
-
-        assert!(!tree.contains(&8));
-    }
-
-    #[test]
-    fn test_btree_remove_from_one_node_tree() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..3 {
-            tree.insert(n);
-        }
-        assert!(tree.contains(&1));
-        assert!(tree.remove(&1));
-        assert!(tree.contains(&0));
-        assert!(!tree.contains(&1));
-        assert!(tree.contains(&2));
-    }
-
-    #[test]
-    fn test_btree_remove_only_key_lower_leaf_in_three_node() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..4 {
-            tree.insert(n);
-        }
-        tree.remove(&3);
-
-        assert!(tree.contains(&0));
-        assert!(tree.remove(&0));
-        assert!(!tree.contains(&0));
-        assert!(tree.contains(&1));
-        assert!(tree.contains(&2));
-    }
-
-    #[test]
-    fn test_btree_remove_only_key_higher_leaf_in_three_node() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in (0..4).into_iter().rev() {
-            tree.insert(n);
-        }
-        tree.remove(&3);
-        assert!(tree.contains(&2));
-        assert!(tree.remove(&2));
-        assert!(tree.contains(&0));
-        assert!(tree.contains(&1));
-        assert!(!tree.contains(&2));
-    }
-
-    #[test]
-    fn test_btree_remove_from_higher_leaf_in_three_node_taking_from_sibling() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in (0..4).into_iter().rev() {
-            tree.insert(n);
-        }
-        assert!(tree.contains(&3));
-        assert!(tree.remove(&3));
-        assert!(!tree.contains(&3));
-    }
-
-    #[test]
-    fn test_btree_remove_from_lower_leaf_in_three_node_taking_from_sibling() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..4 {
-            tree.insert(n);
-        }
-        assert!(tree.contains(&0));
-        assert!(tree.remove(&0));
-        assert!(!tree.contains(&0));
-        assert!(tree.contains(&1));
-        assert!(tree.contains(&2));
-        assert!(tree.contains(&3));
-    }
-
-    #[test]
-    fn test_btree_remove_from_root_in_three_node_causing_merge() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..4 {
-            tree.insert(n);
-        }
-        tree.remove(&3);
-
-        assert!(tree.contains(&1));
-        assert!(tree.remove(&1));
-        assert!(!tree.contains(&1));
-    }
-
-    #[test]
-    fn test_btree_remove_from_root_in_three_node_taking_key_from_higher_child() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in 0..4 {
-            tree.insert(n);
-        }
-        assert!(tree.contains(&1));
-        assert!(tree.remove(&1));
-        assert!(!tree.contains(&1));
-    }
-
-    #[test]
-    fn test_btree_remove_from_root_in_three_node_taking_key_from_lower_child() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        for n in (0..4).into_iter().rev() {
-            tree.insert(n);
-        }
-        assert!(tree.contains(&2));
-        assert!(tree.remove(&2));
-        assert!(!tree.contains(&2));
-    }
-
-    #[test]
-    fn test_btree_iter() {
-        let mut state_builder = TestStateBuilder::new();
-        let mut tree = state_builder.new_btree_set_degree::<2, _>();
-        let keys: Vec<u32> = (0..15).into_iter().collect();
-        for &k in &keys {
-            tree.insert(k);
-        }
-        let iter_keys: Vec<u32> = tree.iter().map(|k| k.clone()).collect();
-        assert_eq!(keys, iter_keys);
-    }
 }
