+++ concordium-std/src/types.rs
@@ -1264,269 +1264,3 @@ pub struct MetadataUrl {
     /// A optional hash of the content.
     pub hash: Option<[u8; 32]>,
 }
-
-/// An ordered map based on [B-Tree](https://en.wikipedia.org/wiki/B-tree), where
-/// each node is stored separately in the low-level key-value store.
-///
-/// It can be seen as an extension adding tracking of the keys ordering on top
-/// of [`StateMap`] to provide functions such as [`higher`](Self::higher) and
-/// [`lower`](Self::lower). This results in some overhead when inserting and
-/// deleting entries from the map compared to using [`StateMap`].
-///
-/// | Operation                                       | Performance   |
-/// |-------------------------------------------------|---------------|
-/// | [`get`](Self::get) / [`get_mut`](Self::get_mut) | O(k)          |
-/// | [`insert`](Self::insert)                        | O(k + log(n)) |
-/// | [`remove`](Self::remove)                        | O(k + log(n)) |
-/// | [`higher`](Self::higher)/[`lower`](Self::lower) | O(k + log(n)) |
-///
-/// Where `k` is the byte size of the serialized keys and `n` is the number of
-/// entries in the map.
-///
-/// ## Type parameters
-///
-/// The map `StateBTreeMap<K, V, S, M>` is parametrized by the types:
-/// - `K`: Keys used in the map. Most operations on the map require this to
-///   implement [`Serialize`](crate::Serialize). Keys cannot contain references
-///   to the low-level state, such as types containing [`StateBox`],
-///   [`StateMap`] and [`StateSet`].
-/// - `V`: Values stored in the map. Most operations on the map require this to
-///   implement [`Serial`](crate::Serial) and
-///   [`DeserialWithState<S>`](crate::DeserialWithState).
-/// - `S`: The low-level state implementation used, this allows for mocking the
-///   state API in unit tests, see
-///   [`TestStateApi`](crate::test_infrastructure::TestStateApi).
-/// - `M`: A `const usize` determining the _minimum degree_ of the B-tree.
-///   _Must_ be a value of `2` or above for the tree to work. This can be used
-///   to tweak the height of the tree vs size of each node in the tree. The
-///   default is set based on benchmarks.
-///
-/// ## Usage
-///
-/// New maps can be constructed using the
-/// [`new_btree_map`][StateBuilder::new_btree_map] method on the
-/// [`StateBuilder`].
-///
-/// ```
-/// # use concordium_std::*;
-/// # use concordium_std::test_infrastructure::*;
-/// # let mut state_builder = TestStateBuilder::new();
-/// /// In an init method:
-/// let mut map1 = state_builder.new_btree_map();
-/// # map1.insert(0u8, 1u8); // Specifies type of map.
-///
-/// # let mut host = TestHost::new((), state_builder);
-/// /// In a receive method:
-/// let mut map2 = host.state_builder().new_btree_map();
-/// # map2.insert(0u16, 1u16);
-/// ```
-///
-/// ### **Caution**
-///
-/// `StateBTreeMap`s must be explicitly deleted when they are no longer needed,
-/// otherwise they will remain in the contract's state, albeit unreachable.
-///
-/// ```no_run
-/// # use concordium_std::*;
-/// struct MyState<S: HasStateApi = StateApi> {
-///     inner: StateBTreeMap<u64, u64, S>,
-/// }
-/// fn incorrect_replace(state_builder: &mut StateBuilder, state: &mut MyState) {
-///     // The following is incorrect. The old value of `inner` is not properly deleted.
-///     // from the state.
-///     state.inner = state_builder.new_btree_map(); // ⚠️
-/// }
-/// ```
-/// Instead, either the map should be [cleared](StateBTreeMap::clear) or
-/// explicitly deleted.
-///
-/// ```no_run
-/// # use concordium_std::*;
-/// # struct MyState<S: HasStateApi = StateApi> {
-/// #    inner: StateBTreeMap<u64, u64, S>
-/// # }
-/// fn correct_replace(state_builder: &mut StateBuilder, state: &mut MyState) {
-///     state.inner.clear_flat();
-/// }
-/// ```
-/// Or alternatively
-/// ```no_run
-/// # use concordium_std::*;
-/// # struct MyState<S: HasStateApi = StateApi> {
-/// #    inner: StateBTreeMap<u64, u64, S>
-/// # }
-/// fn correct_replace(state_builder: &mut StateBuilder, state: &mut MyState) {
-///     let old_map = mem::replace(&mut state.inner, state_builder.new_btree_map());
-///     old_map.delete()
-/// }
-/// ```
-pub struct StateBTreeMap<K, V, S, const M: usize = 8> {
-    pub(crate) map:         StateMap<K, V, S>,
-    pub(crate) ordered_set: StateBTreeSet<K, S, M>,
-}
-
-/// An ordered set based on [B-Tree](https://en.wikipedia.org/wiki/B-tree), where
-/// each node is stored separately in the low-level key-value store.
-///
-/// | Operation                                       | Performance   |
-/// |-------------------------------------------------|---------------|
-/// | [`contains`](Self::contains)                    | O(k + log(n)) |
-/// | [`insert`](Self::insert)                        | O(k + log(n)) |
-/// | [`remove`](Self::remove)                        | O(k + log(n)) |
-/// | [`higher`](Self::higher)/[`lower`](Self::lower) | O(k + log(n)) |
-///
-/// Where `k` is the byte size of the serialized keys and `n` is the number of
-/// entries in the map.
-///
-/// ## Type parameters
-///
-/// The map `StateBTreeSet<K, S, M>` is parametrized by the types:
-/// - `K`: Keys used in the set. Most operations on the set require this to
-///   implement [`Serialize`](crate::Serialize). Keys cannot contain references
-///   to the low-level state, such as types containing [`StateBox`],
-///   [`StateMap`] and [`StateSet`].
-/// - `S`: The low-level state implementation used, this allows for mocking the
-///   state API in unit tests, see
-///   [`TestStateApi`](crate::test_infrastructure::TestStateApi).
-/// - `M`: A `const usize` determining the _minimum degree_ of the B-tree.
-///   _Must_ be a value of `2` or above for the tree to work. This can be used
-///   to tweak the height of the tree vs size of each node in the tree. The
-///   default is set based on benchmarks.
-///
-/// ## Usage
-///
-/// New sets can be constructed using the
-/// [`new_btree_set`][StateBuilder::new_btree_set] method on the
-/// [`StateBuilder`].
-///
-/// ```
-/// # use concordium_std::*;
-/// # use concordium_std::test_infrastructure::*;
-/// # let mut state_builder = TestStateBuilder::new();
-/// /// In an init method:
-/// let mut map1 = state_builder.new_btree_set();
-/// # map1.insert(0u8); // Specifies type of map.
-///
-/// # let mut host = TestHost::new((), state_builder);
-/// /// In a receive method:
-/// let mut map2 = host.state_builder().new_btree_set();
-/// # map2.insert(0u16);
-/// ```
-///
-/// ### **Caution**
-///
-/// `StateBTreeSet`s must be explicitly deleted when they are no longer needed,
-/// otherwise they will remain in the contract's state, albeit unreachable.
-///
-/// ```no_run
-/// # use concordium_std::*;
-/// struct MyState<S: HasStateApi = StateApi> {
-///     inner: StateBTreeSet<u64, S>,
-/// }
-/// fn incorrect_replace(state_builder: &mut StateBuilder, state: &mut MyState) {
-///     // The following is incorrect. The old value of `inner` is not properly deleted.
-///     // from the state.
-///     state.inner = state_builder.new_btree_set(); // ⚠️
-/// }
-/// ```
-/// Instead, the set should be [cleared](StateBTreeSet::clear):
-///
-/// ```no_run
-/// # use concordium_std::*;
-/// # struct MyState<S: HasStateApi = StateApi> {
-/// #    inner: StateBTreeSet<u64, S>
-/// # }
-/// fn correct_replace(state_builder: &mut StateBuilder, state: &mut MyState) {
-///     state.inner.clear();
-/// }
-/// ```
-pub struct StateBTreeSet<K, S, const M: usize = 8> {
-    /// Type marker for the key.
-    pub(crate) _marker_key:  PhantomData<K>,
-    /// The unique prefix to use for this map in the key-value store.
-    pub(crate) prefix:       StateItemPrefix,
-    /// The API for interacting with the low-level state.
-    pub(crate) state_api:    S,
-    /// The ID of the root node of the tree, where None represents the tree is
-    /// empty.
-    pub(crate) root:         Option<state_btree_internals::NodeId>,
-    /// Tracking the number of items in the tree.
-    pub(crate) len:          u32,
-    /// Tracking the next available ID for a new node.
-    pub(crate) next_node_id: state_btree_internals::NodeId,
-}
-
-/// Module with types used internally in [`StateBTreeMap`].
-pub(crate) mod state_btree_internals {
-    use super::*;
-
-    /// Identifier for a node in the tree. Used to construct the key, where this
-    /// node is store in the smart contract key-value store.
-    #[derive(Debug, Copy, Clone)]
-    #[repr(transparent)]
-    pub(crate) struct NodeId {
-        pub(crate) id: u32,
-    }
-
-    /// Type representing a node in the [`StateBTreeMap`].
-    /// Each node is stored separately in the smart contract key-value store.
-    #[derive(Debug)]
-    pub(crate) struct Node<const M: usize, K> {
-        /// List of sorted keys tracked by this node.
-        /// This list should never be empty and contain between `M - 1` and `2M
-        /// - 1` elements.
-        pub(crate) keys:     Vec<K>,
-        /// List of nodes which are children of this node in the tree.
-        ///
-        /// This list is empty when this node is representing a leaf.
-        /// When not a leaf, it will contain exactly `keys.len() + 1` elements.
-        ///
-        /// The elements are ordered such that for a key `keys[i]`:
-        /// - `children[i]` is a subtree containing strictly smaller keys.
-        /// - `children[i + 1]` is a subtree containing strictly larger keys.
-        pub(crate) children: Vec<NodeId>,
-    }
-
-    /// Wrapper implement the exact same deserial as K, but wraps it in an
-    /// option in memory. This is used, to allow taking a key from a mutable
-    /// reference to a node, without cloning the key, during iteration of the
-    /// set.
-    #[repr(transparent)]
-    pub(crate) struct KeyWrapper<K> {
-        pub(crate) key: Option<K>,
-    }
-}
-
-/// An iterator over the entries of a [`StateBTreeSet`].
-///
-/// Ordered by `K`.
-///
-/// This `struct` is created by the [`iter`][StateBTreeSet::iter] method on
-/// [`StateBTreeSet`]. See its documentation for more.
-pub struct StateBTreeSetIter<'a, 'b, K, S, const M: usize> {
-    /// The number of elements left to iterate.
-    pub(crate) length:            usize,
-    /// Reference to a node in the tree to load and iterate before the current
-    /// node.
-    pub(crate) next_node:         Option<state_btree_internals::NodeId>,
-    /// Tracking the nodes depth first, which are currently being iterated.
-    pub(crate) depth_first_stack:
-        Vec<(state_btree_internals::Node<M, state_btree_internals::KeyWrapper<K>>, usize)>,
-    /// Reference to the set, needed for looking up the nodes.
-    pub(crate) tree:              &'a StateBTreeSet<K, S, M>,
-    /// Marker for tracking the lifetime of the key.
-    pub(crate) _marker_lifetime:  PhantomData<&'b K>,
-}
-
-/// An iterator over the entries of a [`StateBTreeMap`].
-///
-/// Ordered by `K`.
-///
-/// This `struct` is created by the [`iter`][StateBTreeMap::iter] method on
-/// [`StateBTreeMap`]. See its documentation for more.
-pub struct StateBTreeMapIter<'a, 'b, K, V, S, const M: usize> {
-    /// Iterator over the keys in the map.
-    pub(crate) key_iter: StateBTreeSetIter<'a, 'b, K, S, M>,
-    /// Reference to the map holding the values.
-    pub(crate) map:      &'a StateMap<K, V, S>,
-}
